+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Fixed
 - Fixed issue 3 in AndersonNRM
 
+### Changed
+- MassAction.__repr__ now also prints rate constant
 
 ## [1.1] - 2018-03-08