+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [1.1.1]
+
+### Fixed
+- Fixed issue 3 in AndersonNRM
+
+### Changed
+- MassAction.__repr__ now also prints rate constant
+
 ## [1.1] - 2018-03-08
 
 ### Fixed